Split out `InstallableFlake` and `InstallableAttrPath`
